move getUsernameFromCric from UserUtilities to ClientUtilities
